Reorganize files and provide documentation
