+++ cmd/init.go
@@ -12,7 +12,7 @@ func executeInit(cmd *cobra.Command, args []string) {
 	var err error
 
 	// Create pizza directory
-	project, err := mraziz.GetBaseProject()
+	project, err := mraziz.GetBasePizza()
 	if err != nil {
 		panic(err)
 	}
@@ -34,14 +34,4 @@ var initCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(initCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// initCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// initCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
