--- mraziz/generatable.go
@@ -1,5 +1,0 @@
-package mraziz
-
-type Generatable interface {
-	Generate() error
-}
